requirements.txt
streamlit>=1.25.0
pandas
toml
python-dotenv
openai
